Move crates lowering into seperate module
